Most changes were made in the function-braindump-stuff notebook. All important/relevant coding was also moved into that notebook.
A new version of the squared_diff function (temp name sd2) was written to be able to take in vector arguments, esp the vector for 2015 observed mmr. This is in conjunction w/ a rewritten version of the sdg_mmr_2030 function (SM2030) that takes in that 2015 mmr vector as an argument and outputs a single vector of 2030 mmr sdg predictions.

Note that these rewritten functions will likely replace the old ones due to better functionality in conjunction w/ e/o.