Fixed the recoded sdg_mmr2030 function because it was incorrect (missing the adjusted mmr to make any countries w/ high mmr have a 140 mmr goal). I had to make several changes to the other recoded functions as a result.
Organized the recoded functions in a way that makes more logical sense and removed unnecessary code and commented questions that I have since answered for myself.

Fixed some function arguments and inner temporary variables to have more intuitive names and just to make the function run better overall.

Started codin the arr_sdg per country function (but unsure if the numbers returned are reasonable so put on hold for now).